Create nameservers in production AWS account for subdomains
The TLD hosted zone is in the productin AWS account. To support non-live
environment sub domains, a NS record needs to be created in the
production account for each of the non-live environments (development,
pre-production).

Automate this by running a script on the production build pipeline,
which will:
1. Get the nameservers for dev, prep
2. Create an upsert json payload
3. Apply this to production route53 using the aws cli

This will be idempotent and can be run for each build, if the records
exist, it will be a noop.