Update Radius cluster auto scaling
Instead of monitoring CPU and memory of the cluster, which doesn't
correlate directly with the amount of work that FreeRadius is doing,
look at the number of packets processed per container instead.

This gives us fine grained control over how much traffic/authentications we'd
like each container to process.

Currently set to 60 packets per container to test this scaling event, an
investigation needs to be done to find out what this number will be on
the production service. TBC during performance testing.