Re-instate geo-location routing policy for admin CNAME
This was suspected to have caused issues with intermittent DNS
resolution of the admin portal. The cause was found to be missing
hosted zones in the parent AWS account.

We want to restrict admin URL resolution to the UK. Remains to be seen
if this reduces the number of unauthorised login attempts. We don't
expect administrators to be configuring NACS from outside the UK.