Silence required provider warning message
In the new versions of Terraform, required_providers have to be
explicity defined in the module using it.

Currently we get the following message whenever applying:

`Warning: Provider aws is undefined

  on main.tf line 110, in module "radius":
 110:     aws = aws.env

Module module.radius does **t declare a provider named aws.
If you wish to specify a provider configuration for the module, add an entry
for aws in the required_providers block within the module.

(and 10 more similar warnings elsewhere)`

Fix this by adding required_providers to each module.