:wrench: Update resource to account for complex queries
This commit updates the aws_cloudwatch_log_metric_filter resource to account for the scenario where the log_filters list in main contains entries with queries. The first resource handles 'normal' elements of the list and the latter resource is setup to filter for entries containing '?' and create metrics based off of this.