Amended Docker run and gen-env script
An issue in which the environment variables which were not explicitly TF_VAR_ or ENV would not be available to the Docker container's environment when running was discovered.

By changing how we source the whole .env file and run various commands the environment variables are fully availible for the environemt for both terraform and scripting purposes.