Align ECR retention policy with agreed policy
The current policy of
"Expire images older than 14 days"
Doesn't match our agreed policy as used in DNS & DHCP services.
This also creates a risk that we can't rollback to previous known good images as any image older than 14 days will be expired.

This is actually hampering ability to diagnose a possible issue with a new Radius AMI.

The new policy has been agreed with the balance of retention and ability to rollback to known  good images that might be many months old.
"Expire older versions of untagged images, keeping the latest 15"

ND-463