Added new scripts for the DB access
Add connection files to gitignore we don't want to commit those.

Created new Make targets for DB connections
Make it really easy for an engineer to get the information for connecting to the RDS DB from the bastion when debugging issues.

Update the documentation for RDS Bastion
reflect new changes since AWS Secrets have been used and the explain to use the new make targets.
ND-510